@@ -1,5 +1,6 @@
 import 'dart:ffi';
 import 'package:flutter/material.dart';
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 
 const double bottomContainerHeight = 80.0;
 const Color activeCardColour = Color(0xFF1D1E33);
@@ -24,25 +25,46 @@ class _InputPageState extends State<InputPage> {
               child: Row(
                 children: <Widget>[
                   Expanded(
-                    child: ReusableCard(activeCardColour),
+                    child: ReusableCard(
+                      colour: activeCardColour,
+                      cardChild: Column(
+                        mainAxisAlignment: MainAxisAlignment.center,
+                        children: <Widget>[
+                          Icon(
+                            FontAwesomeIcons.mars,
+                            size: 80.0,
+                          ),
+                          SizedBox(
+                            height: 15.0,
+                          ),
+                          Text(
+                            'MALE',
+                            style: TextStyle(
+                              fontSize: 18.0,
+                              color: Color(0xFF8D8E98),
+                            ),
+                          )
+                        ],
+                      ),
+                    ),
                   ),
                   Expanded(
-                    child: ReusableCard(activeCardColour),
+                    child: ReusableCard(colour: activeCardColour),
                   ),
                 ],
               ),
             ),
             Expanded(
-              child: ReusableCard(activeCardColour),
+              child: ReusableCard(colour: activeCardColour),
             ),
             Expanded(
                 child: Row(
               children: <Widget>[
                 Expanded(
-                  child: ReusableCard(activeCardColour),
+                  child: ReusableCard(colour: activeCardColour),
                 ),
                 Expanded(
-                  child: ReusableCard(activeCardColour),
+                  child: ReusableCard(colour: activeCardColour),
                 ),
               ],
             )),
@@ -61,12 +83,14 @@ class _InputPageState extends State<InputPage> {
 
 class ReusableCard extends StatelessWidget {
   final Color colour;
+  final Widget? cardChild;
 
-  ReusableCard(@required this.colour);
+  ReusableCard({required this.colour, this.cardChild});
 
   @override
   Widget build(BuildContext context) {
     return Container(
+      child: cardChild,
       margin: EdgeInsets.all(15.0),
       decoration: BoxDecoration(
         borderRadius: BorderRadius.circular(10.0),
